Public version of the NUMA ACPI 6.3 White Paper.
This version is ready to share with the open source community.

I've squashed own all of the commits to present a single coherent
base to move forward from.

Given uncertainty around this being an official UEFI white paper
I've commented out the UEFI logo include and trademark statement
until that is resolved.

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>